Flash/bounce taskbar/dock icon more often, but for less time
In our world Mattermost is pretty low traffic, so if someone is talking
I'd at least like the dock icon to bounce to get my attention so call
notifyOS for both unread channels and mentions.  To avoid the icon
bouncing forever, downgrade the bounce request from 'critical' to
'informational' so MacOS will stop the bounce after about a second.

In the future this should be a pair of user configurable properties.