Add build improvements from upstream while avoiding wine dependency
The Matterfront project has added some nice updates (capitalize the
application name, add an icon, etc.) but trying to build the latest is a
bit of a gamble right now (there's a weird off-by-one notification bug,
and there's a lot of churn currently).  So just backport some of those
niceties to our current GenArts version.

Unfortunately for Electron Packager to set the application icon for
Windows it needs wine.  That's a pretty heavy dependency for something
so simple, so I split the build and just updated the icon using Resource
Hacker on Windows.